changes folder structure, adds a skip button, volume slider, and toggle play
