+++ package.json
@@ -1,8 +1,52 @@
 {
-  "name": "feedplayer",
-  "version": "1.0.0",
-  "main": "dist/index.js",
-  "repository": "git@github.com:platformfirst/feedfm-player-exploration.git",
-  "author": "abhiaiyer91 <abhiaiyer91@gmail.com>",
-  "license": "MIT"
+  "name": "gatsby-starter-default",
+  "private": true,
+  "description": "A simple starter to get up and developing quickly with Gatsby",
+  "version": "0.1.0",
+  "author": "Kyle Mathews <mathews.kyle@gmail.com>",
+  "dependencies": {
+    "@chakra-ui/core": "^0.8.0",
+    "@emotion/core": "^10.0.35",
+    "@emotion/styled": "^10.0.27",
+    "@hot-loader/react-dom": "16.12.0",
+    "emotion-theming": "^10.0.27",
+    "feed-media-audio-player": "^1.99.30",
+    "gatsby": "^2.24.66",
+    "gatsby-image": "^2.4.20",
+    "gatsby-plugin-chakra-ui": "^0.1.4",
+    "gatsby-plugin-google-fonts": "^1.0.1",
+    "gatsby-plugin-manifest": "^2.4.32",
+    "gatsby-plugin-offline": "^3.2.29",
+    "gatsby-plugin-react-helmet": "^3.3.12",
+    "gatsby-plugin-sharp": "^2.6.37",
+    "gatsby-source-filesystem": "^2.3.31",
+    "gatsby-transformer-sharp": "^2.5.16",
+    "prop-types": "^15.7.2",
+    "react": "^16.12.0",
+    "react-dom": "16.12.0",
+    "react-helmet": "^6.1.0"
+  },
+  "devDependencies": {
+    "prettier": "2.1.2"
+  },
+  "keywords": [
+    "gatsby"
+  ],
+  "license": "0BSD",
+  "scripts": {
+    "build": "gatsby build",
+    "develop": "gatsby develop",
+    "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
+    "start": "npm run develop",
+    "serve": "gatsby serve",
+    "clean": "gatsby clean",
+    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/gatsbyjs/gatsby-starter-default"
+  },
+  "bugs": {
+    "url": "https://github.com/gatsbyjs/gatsby/issues"
+  }
 }
