+++ src/gatsby-plugin-chakra-ui/theme.js
